uncouple flask utils functions from utils module
